Revert "Bump gradle from 3.5.3 to 4.0.0 in /mobile/android"
As commented out in [1], for now, a workaround is to explicitly set the
NDK version in the build.gradle file. However, this is not easily
do-able, since local and CI environments do not necessarily leverage the
same NDK.

So for now, we are skipping this version bump of the Android Gradle
Plugin, but are keeping the other commits.

This reverts commit 25f726f363beb474fcfed9e9bbbbbf221de90fbc.